[WIP] Bitbucket API SCM resolution
On a plane, so testing this is really hard and it would suck if a "planefactor"
issue made this code disappear. :P -- Will amend this commit message to link to
the related issue once things are looking good (in terms of CI).

This change introduces a new SCM resolution mechanism for Bitbucket repos, to
hopefully avoid false-positives when reaching out to Bitbucket to get the repo's
metadata. Bitbucket is a weird code repository solution in that it supports both
`git` and `hg` (mercurial) for hosting repos. This makes checking out source
code from Bitbucket less straightforward than other places.

Previously the deducer tried to make educated guesses around which SCM was in
use based on a few things, largely hints from the package's import path. It also
had some fallbacks in place in case it got it wrong, but it did still cause some
issues in a few cases when it got it wrong.

This change introduces a minimal internal Bitbucket API client (`bitbucketapi`)
with the single purpose of reaching out to Bitbucket to get the SCM type for the
repository. This deprecates the need for the earlier heuristics, and should
result in a correct result every time. This client was written to be consumed as
a singleton (calling functions on the package) to avoid complicating the
`bitbucketDeducer` further.

The client has no caching of previously seen SCM results for a package, which
could result in more API calls than we'd like. One optimization would be to keep
a map of previously resolved packages, and to serve the result from there.

Tests are also implemented for the `bitbucketapi` package, hitting `97.0%` code
coverage.

Signed-off-by: Tim Heckman <t@heckman.io>